Support fragments in links
Separately parse the `#`-portion of a reference in a wikilink so that
references like the following are valid and easy to process in the
resolver.

    [[Foo#Bar]]
    [[#Foo]]

The latter is important to make references to headers in the same
document.

These are still passed to the resolver to manipulate as needed.
This is necessary because some wikilink formats allow these to be full
names of headers (`[[#Some section]]`) while others prefer IDs of those
headers (`[[#some-section]]`).