render: Fix alt text handling
Embedded nodes that are images should not render their child node
because the child node is the alt text/label.
Instead, we should extract the label and turn it into alt text.

This has a special case: the alt text is not rendered
if the target and the alt text are the same.
This prevents `[[foo.jpg]]` from using `alt="foo.jpg"`.